third_party/blink/web_tests/fast/css/counters/update-reset-counter-parent-sibling-expected.txt

This is a testharness.js-based test.
FAIL Dynamically changing counter-reset with counter-resets in parent and sibling updates counters accordingly assert_equals: expected "1" but got "2"
Harness: the test ran to completion.
